refactor: using cobra cli
